move fastq and bam validation to subfolder
