+++ src/grz_cli/validation/fastq.py
@@ -13,8 +13,8 @@ from typing import TextIO
 
 from tqdm.auto import tqdm
 
-from .constants import TQDM_SMOOTHING
-from .file_operations import TqdmIOWrapper
+from ..constants import TQDM_SMOOTHING
+from ..file_operations import TqdmIOWrapper
 
 log = logging.getLogger(__name__)
